+++ R/bsmote.R
@@ -11,8 +11,6 @@
 #'  for more details. The selection should result in _single
 #'  factor variable_. For the `tidy` method, these are not
 #'  currently used.
-#' @param role Not used by this step since no new variables are
-#'  created.
 #' @param column A character string of the variable name that will
 #'  be populated (eventually) by the `...` selectors.
 #' @param neighbors An integer. Number of nearest neighbor that are used
@@ -152,11 +150,13 @@ step_bsmote <-
     over_ratio = 1,
     neighbors = 5,
     all_neighbors = FALSE,
+    indicator_column = NULL,
     skip = TRUE,
     seed = sample.int(10^5, 1),
     id = rand_id("bsmote")
   ) {
     check_number_whole(seed)
+    check_string(indicator_column, allow_null = TRUE, allow_empty = FALSE)
 
     add_step(
       recipe,
@@ -169,6 +169,7 @@ step_bsmote <-
         neighbors = neighbors,
         all_neighbors = all_neighbors,
         predictors = NULL,
+        indicator_column = indicator_column,
         skip = skip,
         seed = seed,
         id = id
@@ -186,6 +187,7 @@ step_bsmote_new <-
     neighbors,
     all_neighbors,
     predictors,
+    indicator_column,
     skip,
     seed,
     id
@@ -200,6 +202,7 @@ step_bsmote_new <-
       neighbors = neighbors,
       all_neighbors = all_neighbors,
       predictors = predictors,
+      indicator_column = indicator_column,
       skip = skip,
       seed = seed,
       id = id
@@ -217,6 +220,13 @@ prep.step_bsmote <- function(x, training, info = NULL, ...) {
   check_1_selected(col_name)
   check_column_factor(training, col_name)
 
+  recipes::check_name(
+    tibble(x = logical(0)),
+    training,
+    x,
+    newname = x$indicator_column
+  )
+
   predictors <- setdiff(recipes::recipes_names_predictors(info), col_name)
 
   check_type(training[, predictors], types = c("double", "integer"))
@@ -231,6 +241,7 @@ prep.step_bsmote <- function(x, training, info = NULL, ...) {
     neighbors = x$neighbors,
     all_neighbors = x$all_neighbors,
     predictors = predictors,
+    indicator_column = x$indicator_column,
     skip = x$skip,
     seed = x$seed,
     id = x$id
@@ -251,6 +262,7 @@ bake.step_bsmote <- function(object, new_data, ...) {
     return(new_data)
   }
 
+  n_orig <- nrow(new_data)
   new_data <- as.data.frame(new_data)
 
   predictor_data <- new_data[, col_names]
@@ -269,6 +281,8 @@ bake.step_bsmote <- function(object, new_data, ...) {
     }
   )
   new_data <- na_splice(new_data, synthetic_data, object)
+
+  new_data <- add_indicator_column(new_data, n_orig, object$indicator_column)
 
   new_data
 }